# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexa AI SaaS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface UserRegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface UserLoginInput {
  email: string;
  password: string;
}

export type AuthUserPlan = typeof AuthUserPlan[keyof typeof AuthUserPlan];


export const AuthUserPlan = {
  free: 'free',
  pro: 'pro',
} as const;

export interface AuthUser {
  id: number;
  name: string;
  email: string;
  plan: AuthUserPlan;
  dailyMessagesUsed: number;
  dailyLimit: number;
  createdAt: string;
}

export type UserProfilePlan = typeof UserProfilePlan[keyof typeof UserProfilePlan];


export const UserProfilePlan = {
  free: 'free',
  pro: 'pro',
} as const;

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  plan: UserProfilePlan;
  dailyMessagesUsed: number;
  dailyLimit: number;
  totalConversations: number;
  totalMessages: number;
  createdAt: string;
}

export interface Conversation {
  id: number;
  title: string;
  messageCount?: number;
  createdAt: string;
  updatedAt: string;
}

export interface ConversationInput {
  /** @minLength 1 */
  title: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface ChatResponse {
  userMessage: Message;
  assistantMessage: Message;
  dailyMessagesUsed: number;
  dailyLimit: number;
}

export interface Plan {
  id: string;
  name: string;
  price: number;
  currency: string;
  features: string[];
  dailyLimit: number;
  recommended?: boolean;
}

export interface PaymentPreferenceInput {
  planId: string;
}

export interface PaymentPreference {
  id: string;
  initPoint: string;
  sandboxInitPoint: string;
}

export type WebhookPayloadData = {
  id?: string;
};

export interface WebhookPayload {
  type?: string;
  data?: WebhookPayloadData;
}