# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexa AI SaaS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from './message';

export interface ChatResponse {
  userMessage: Message;
  assistantMessage: Message;
  dailyMessagesUsed: number;
  dailyLimit: number;
}