# lib/api-zod/src/generated/types/paymentPreference.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexa AI SaaS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentPreference {
  id: string;
  initPoint: string;
  sandboxInitPoint: string;
}