# lib/api-zod/src/generated/types/paymentPreferenceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexa AI SaaS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentPreferenceInput {
  planId: string;
}