# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexa AI SaaS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Plan {
  id: string;
  name: string;
  price: number;
  currency: string;
  features: string[];
  dailyLimit: number;
  recommended?: boolean;
}